Move Plugin type to common from schemas
