+++ packages/generator/src/types.ts
@@ -1,12 +1,10 @@
-import { Schema } from "@pdfme/common"
-import { PDFRenderProps } from '@pdfme/schemas';
+import { PDFRenderProps } from '@pdfme/common';
 
 export type EmbedPdfBox = {
   mediaBox: { x: number; y: number; width: number; height: number };
   bleedBox: { x: number; y: number; width: number; height: number };
   trimBox: { x: number; y: number; width: number; height: number };
 };
-
-export interface Renderer<T extends Schema = any> {
-  [key: string]: { render: (arg: PDFRenderProps<T>) => Promise<void> } | undefined;
+export interface Renderer { // FIXME anyをなくす
+  [key: string]: { render: (arg: PDFRenderProps<any>) => Promise<void> } | undefined;
 }
