+++ packages/schemas/src/barcodes/index.ts
@@ -2,8 +2,7 @@ import { pdfRender } from './pdfRender';
 import { getPropPanelByBarcodeType } from './propPanel';
 import { uiRender } from './uiRender';
 import type { BarcodeSchema } from './types';
-import type { Plugin } from '../types';
-import { barcodeSchemaTypes } from '@pdfme/common';
+import { Plugin,barcodeSchemaTypes } from '@pdfme/common';
 
 const schemas = barcodeSchemaTypes.reduce((acc, type) => Object.assign(acc, {
     [type]: {
