+++ packages/schemas/src/text/index.ts
@@ -1,8 +1,8 @@
+import type { Plugin } from '@pdfme/common';
 import { pdfRender } from './pdfRender';
 import { propPanel } from './propPanel';
 import { uiRender } from './uiRender';
 import type { TextSchema } from './types';
-import type { Plugin } from '../types';
 
 const schema: Plugin<TextSchema> = {
     pdf: pdfRender,
